requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
djangorestframework>=3.14.0
django-filter>=23.0
drf-spectacular>=0.26.0